requirements.txt
PyQt6==6.4.0
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.1
supabase>=2.0.0
gotrue>=2.0.0
httpx>=0.28.1
Pillow==10.2.0
pytest==8.0.0
python-dotenv==0.19.0
PyInstaller==6.3.0